Move useAsyncRender to web file
