+++ mod.web.ts
@@ -4,3 +4,4 @@ export {
   proxy,
   withManifest,
 } from "./clients/withManifest.ts";
+export { clickOnAsyncRenderElements } from "./hooks/client/useAsyncRender.ts";
